solve a bug

+++ src/io/port.h
@@ -12,7 +12,7 @@ uint16_t inw(uint16_t port);
 uint32_t inl(uint16_t port);
 
 void insb(uint16_t port, uint8_t *buf, size_t count);
-void insw(uint16_t port, uint8_t *buf, size_t count);
-void insl(uint16_t port, uint8_t *buf, size_t count);
+void insw(uint16_t port, uint16_t *buf, size_t count);
+void insl(uint16_t port, uint32_t *buf, size_t count);
 
 #endif
